Add attach_dataset fixture in Commons malformed auth test
